+++ src/index.ts
@@ -1,7 +1,7 @@
-export { WindowPostMessageStream } from './WindowPostMessageStream';
-export { WorkerPostMessageStream } from './WorkerPostMessageStream';
-export { WorkerParentPostMessageStream } from './WorkerParentPostMessageStream';
-export { ParentProcessMessageStream } from './ParentProcessMessageStream';
-export { ChildProcessMessageStream } from './ChildProcessMessageStream';
+export { WindowPostMessageStream } from './window/WindowPostMessageStream';
+export { WorkerPostMessageStream } from './WebWorker/WorkerPostMessageStream';
+export { WorkerParentPostMessageStream } from './WebWorker/WorkerParentPostMessageStream';
+export { ParentProcessMessageStream } from './node-process/ParentProcessMessageStream';
+export { ChildProcessMessageStream } from './node-process/ChildProcessMessageStream';
 export { PostMessageEvent } from './BasePostMessageStream';
 export { StreamData, StreamMessage } from './utils';
